Add support for splitting log files
By default, FileLogger will now create new log files when they
are more than 1MB. This can be customized with the 'fileSizeCap'-
property.

Also, you can set 'shouldKeepArchivedLogs' to 'false' to just
delete old logs when reaching a certain size. This has the drawback
of losing all current logs and starting fresh when reaching the cap,
instead of just deleting entries in the beginning of the file.